Drawing of effective sequence (#122)
* Drawing of effective sequence

* corrected style

* small corrections

* style

* proper plot

* Fixing time plot data monotonicity

* cubic spline interpolation

wrote down a version without using explicitely gather_data, but I guess it should be rewritten with it, once the time ordering issue is resolved

* Corrections for ubuntu tests?

* 100% coverage

added simulation drawing test for constant waveforms

* Sim drawer + sampling time/evaluation time distinction

The simulation class now has two attributes _times and evaluation_times corresponding respectively to the one used in the creation of the Hamiltonian and the one given explicitely to sesolve, where it is suppose to evaluate the state

* Index offset correction

Corrected the computation of the cubic spline interpolation points

* Final corrections

When the pulse is constant-by-parts, there are some artifacts at the jumps, due to the spline interpolation done by sesolve. To circumvent this problem, there is an option in the QobjEvo contruction to specify that the coefficients are piecewise constant. No interpolation is done in this case

* Update pulser/_seq_drawer.py

Co-authored-by: Henrique Silvério <henrique.silverio@tecnico.ulisboa.pt>

* Added suggested corrections

* Code commenting

Added explanation of confusing lines  in seq_drawer

* Resolving conflicts

* Resolved conflicts

* Delete simulation.py

* Update simulation.py

corrected the ns vs μs incoherence in the definition of simulation_times

* Update pulser/simulation/simulation.py

Co-authored-by: Henrique Silvério <henrique.silverio@tecnico.ulisboa.pt>

* resolving nits

implemented suggestions

Co-authored-by: Henrique Silvério <henrique.silverio@tecnico.ulisboa.pt>