Improve test and get total coverage
Table based tests

Add a modulation test

I don't like it. It is testing the sampling for sure, but still I would
like to test against a known output of the modulation.

Remove unnecessary if branch

functools.reduce got us covered already

Extend to digital

Fix mypy error because of weird scope of variables

Test sequence in XY and fix a typo in the relevant code

Test for corner cases and omit defensive coding with pragma no cover