Add a note about performance for sampling of global channels
For global channels, we hold the same data in N copies, N the number of
qubits of the register. It scales poorly with the size of the register,
but it remains manageable for current size of registers.

If needed, we should patch this, at the expense of a more complex
control flow in the sample() function at least.